+++ MiDAS/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 0.0.0.9034 - 1/10/2019
++ add function for models comparison using likelihood ratio test (LRTest)
++ add function for performing omnibus test on amino acid positions (aaPosOmnibusTest)
+
+## 0.0.0.9034 - 1/10/2019
 + in summariseAAPosition output alleles numbers are now sorted alphabeticaly
 
 ## 0.0.0.9032 - 21/09/2019